chore: :truck: rename files and functions to match naming scheme
