+++ sprout/views/__init__.py
@@ -7,7 +7,7 @@ from django.shortcuts import render
 # https://simpleisbetterthancomplex.com/tutorial/2016/08/02/how-to-split-views-into-multiple-files.html
 from .column_review import column_review
 from .data_import import data_import
-from .project_id_data_id_upload import project_id_data_id_upload
+from .project_id_metadata_id_data_update import project_id_metadata_id_data_update
 from .metadata_create import metadata_create
 from .table_files import table_file_download, table_files
